+++ src/auth/resourceServer.js
@@ -32,6 +32,7 @@ const {
 } = process.env
 const hubs = hub.split(',')
 const emailCheck = require('email-validator')
+const { USER_ROLES } = require('./consts')
 const handleCheck = '^[A-Za-z0-9-]{3,32}$'
 const nameCheck = '^[A-Za-z0-9_~ -]{3,32}$'
 
@@ -47,6 +48,7 @@ module.exports = {
   open: [passport.authenticate(['bearer', 'anonymous'], { session: false }), cors()],
   /** auth for OAuth client / service account jwt login (e.g. in Authorization code grant) */
   clnt: passport.authenticate('oauth2-client-jwt', { session: false }),
+  admn: [passport.authenticate('bearer', { session: false }), requireAdmin],
   scope,
   /** Require access to view private information */
   viewScope: scope(scopes.viewPrivate.name),
@@ -56,6 +58,7 @@ module.exports = {
   localToken: [hubCors, localToken],
   /** terminate login session */
   logout: [hubCors, logout],
+  oidcLoginProviders,
   /** for endpoints that behave differently for authorized requests */
   passIfNotAuthorized,
   requirePrivilege,
@@ -68,7 +71,8 @@ module.exports = {
     returnTo
   ],
   validateNewUser,
-  registration: [registerUser, respondRedirect]
+  returnTo,
+  respondRedirect
 }
 
 /// side effects ///
@@ -189,6 +193,13 @@ function dynamicCorsFromToken (req, done) {
   } catch (err) { done(err) }
 }
 
+/** List OpenId providers that should have custom login buttons */
+function oidcLoginProviders (req, res, next) {
+  authdb.getOidcLoginProviders().then(providers => {
+    res.json(providers)
+  }).catch(next)
+}
+
 /** for endpoints that behave differently for authorized requests */
 function passIfNotAuthorized (req, res, next) {
   if (!req.get('authorization')) {
@@ -241,8 +252,8 @@ function userToActor (req, res, next) {
 
 async function registerUser (req, res, next) {
   try {
-    const { username, password, email } = req.body
-    const user = await authdb.createUser(username, password, email)
+    const { username, password, email, oidcProviders } = req.body
+    const user = await authdb.createUser(username, password, email, oidcProviders)
     if (!user) {
       throw new Error('Unable to create user')
     }
@@ -266,50 +277,53 @@ function changePassword (req, res, next) {
 }
 
 async function validateNewUser (req, res, next) {
-  // check validity
-  let validMessage = ''
-  if (!emailCheck.validate(req.body.email)) {
-    validMessage += 'Invalid email. '
-  } else {
-    req.body.email = req.body.email.toLowerCase()
-  }
+  try {
+    // check validity
+    let validMessage = ''
+    if (!emailCheck.validate(req.body.email)) {
+      validMessage += 'Invalid email. '
+    } else {
+      req.body.email = req.body.email.toLowerCase()
+    }
 
-  if (!RegExp(handleCheck).test(req.body.username)) {
-    validMessage += `Username must match ${handleCheck}. `
-  } else {
-    req.body.username = req.body.username.toLowerCase()
-  }
-  if (req.body.name && !RegExp(nameCheck).test(req.body.name)) {
-    validMessage += `Display name must match ${nameCheck}. `
-  } else if (!req.body.name) {
-    // display name is optional in registration, default to username
-    req.body.name = req.body.username
-  }
+    if (!RegExp(handleCheck).test(req.body.username)) {
+      validMessage += `Username must match ${handleCheck}. `
+    } else {
+      req.body.username = req.body.username.toLowerCase()
+    }
+    if (req.body.name && !RegExp(nameCheck).test(req.body.name)) {
+      validMessage += `Display name must match ${nameCheck}. `
+    } else if (!req.body.name) {
+      // display name is optional in registration, default to username
+      req.body.name = req.body.username
+    }
 
-  if (validMessage) {
-    return res.status(400).format({
-      text: () => res.send(validMessage),
-      json: () => res.json({ error: validMessage })
-    })
-  }
-  // check availability
-  let availableMessage = ''
-  if (await authdb.getUserByEmail(req.body.email)) {
-    availableMessage += 'Email already registered. '
-  }
+    if (validMessage) {
+      return res.status(400).format({
+        text: () => res.send(validMessage),
+        json: () => res.json({ error: validMessage })
+      })
+    }
+    // check availability
+    let availableMessage = ''
+    if (await authdb.getUserByEmail(req.body.email)) {
+      availableMessage += 'Email already registered. '
+    }
 
-  if (await authdb.getUserByName(req.body.username)) {
-    availableMessage += 'Username already registered. '
-  }
+    if (await authdb.getUserByName(req.body.username)) {
+      availableMessage += 'Username already registered. '
+    }
 
-  if (availableMessage) {
-    return res.status(409).format({
-      text: () => res.send(availableMessage),
-      json: () => res.json({ error: availableMessage, taken: true })
-    })
+    if (availableMessage) {
+      return res.status(409).format({
+        text: () => res.send(availableMessage),
+        json: () => res.json({ error: availableMessage, taken: true })
+      })
+    }
+    next()
+  } catch (err) {
+    next(err)
   }
-
-  next()
 }
 
 function respondRedirect (req, res) {
@@ -328,4 +342,13 @@ function returnTo (req, res) {
     delete req.session.returnTo
   }
   res.redirect(redirect)
+}
+
+function requireAdmin (req, res, next) {
+  const isAdmin = req.user?.role === USER_ROLES.ADMIN
+  if (isAdmin) {
+    next()
+  } else {
+    return res.sendStatus(403)
+  }
 }